// frontend/src/types/api/player-stats.ts
import { GameDetailsResponse } from './game';
import { PlayerBoxscoreResponse, PlayerResponse } from './player';
import { TeamDetailsResponse } from './team';

export type PlayerStatsResponse = {
	id: number;
	documentId: string;
	game: GameDetailsResponse;
	team: TeamDetailsResponse;
	player: PlayerResponse;
	status: 'starter' | 'bench' | 'dnp-cd';
	playerNumber: string;
	isCaptain: boolean;
	minutes: string;
	seconds: string;
	points: string;
	fieldGoalsMade: string;
	fieldGoalsAttempted: string;
	threePointersMade: string;
	threePointersAttempted: string;
	freeThrowsMade: string;
	freeThrowsAttempted: string;
	rebounds: string;
	offensiveRebounds: string;
	defensiveRebounds: string;
	assists: string;
	steals: string;
	blocks: string;
	turnovers: string;
	fouls: string;
	foulsOn: string;
	blocksReceived: string;
	plusMinus: string;
	efficiency: string;
	createdAt: string;
};

export type PlayerStatRankings = {
	points_rank: number;
	steals_rank: number;
	assists_rank: number;
	rebounds_rank: number;
	offensive_rebounds_rank: number;
	defensive_rebounds_rank: number;
	blocks_rank: number;
	turnovers_rank: number;
	fouls_rank: number;
	fouls_on_rank: number;
	blocks_received_rank: number;
	plus_minus_rank: number;
	efficiency_rank: number;
	def_rebounds_rank: number;
	off_rebounds_rank: number;
	field_goals_made_rank: number;
	field_goals_attempted_rank: number;
	three_pointers_made_rank: number;
	three_pointers_attempted_rank: number;
	free_throws_made_rank: number;
	free_throws_attempted_rank: number;
};

export type PlayerRecords = PlayerBoxscoreResponse & PlayerStatRankings;

export type PlayerStatsFormData = {
	gameId: number;
	teamId: number;
	playerId: number;
	status: 'starter' | 'bench' | 'dnp-cd';
	isCaptain: boolean;
	playerNumber: string;
	minutes: string;
	seconds: string;
	points: string;
	fieldGoalsMade: string;
	fieldGoalsAttempted: string;
	threePointersMade: string;
	threePointersAttempted: string;
	freeThrowsMade: string;
	freeThrowsAttempted: string;
	rebounds: string;
	offensiveRebounds: string;
	defensiveRebounds: string;
	assists: string;
	steals: string;
	blocks: string;
	turnovers: string;
	fouls: string;
	foulsOn: string;
	blocksReceived: string;
	plusMinus: string;
	efficiency: string;
};